requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
black==22.6.0
chardet==4.0.0
click==8.1.3
colorama==0.4.5
cowsay==5.0
discord.py==1.7.3
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
python-dotenv==0.20.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
yarl==1.7.2
